fix(export): fail closed on a geometry-less COLLADA/KMZ export
`write_dae` emits `<library_effects>`, `<library_materials>`,
`<library_geometries>` and a `<visual_scene>` unconditionally. When no
triangle survives, all four come out with no children, which COLLADA 1.4.1
forbids - each of those children is `minOccurs="1"` - so the document does
not validate and Google Earth loads nothing from it.
`export_kmz_collada_from_meshes` returns `Vec<u8>` and therefore has no
error channel at all, so it packed that document and every caller read the
result as success. Measured: 1590 bytes from zero meshes, no error.

Two triggering inputs, and the second is the one a mesh-count check cannot
see:

  1. an empty visible set;
  2. a mesh whose only triangle is degenerate. The vertex dedup collapses
     two of its corners and the triangle is then dropped for having two
     equal indices - AFTER its material was registered - so effects and
     materials are NON-EMPTY while geometries is empty.

`try_export_collada_from_meshes` and `try_export_kmz_collada_from_meshes`
return `ExportError::NoRenderGeometry` instead, and the wasm
`exportKmzFromMeshes` boundary now calls the fallible one, so the viewer's
KMZ export stops silently shipping a file Google Earth cannot open. The
GLB twin already had this sibling and documents the same reasoning
(`gltf/from_meshes.rs`, #1438/#1516). Both viewer call sites already catch
(`KmzExportDialog.handleExport`, `LocationMap.handleExportKmz`), so the
throw surfaces as a toast rather than an unhandled rejection.

The emptiness check reads the material buckets AFTER every drop rule has
run, not the caller's `index_counts`, so trigger 2 reaches it, and it runs
before the AABB centring so the failure path does no wasted work.

`/code-review`: the same loop `break`ed silently on a declared count
running past its buffer, so the input the GLB twin refuses with
`MalformedMeshInput` came back `Ok` here with every later mesh missing.
It now returns the same error, named for the mesh, and the wasm
`exportKmzFromMeshes` doc and the changeset say so. `export_kmz` (GLB in
KMZ, which Google Earth cannot load) is superseded by its own account and
stays only because deleting it is a wasm/bridge/index.ts surface removal;
it carries a `TODO(remove-by: #4591)`.

SPLIT, not a budget raise: `collada.rs` was allowlisted at 430 lines and
the new entry point took it to 470. The COLLADA document WRITER moved
verbatim to `collada_document.rs`, bringing `collada.rs` well under 400, so the
allowlist row is DELETED rather than raised and the allowlist total goes
down. `rust/export`'s `ALLOWLIST_DIGESTS` pin moves with it. `kmz.rs`'s
inline `#[cfg(test)]` block moved to `kmz_tests.rs` under the house
pattern, keeping that file (unallowlisted) under 400.

`packages/wasm/pkg/ifc-lite.d.ts` is regenerated with
`scripts/build-wasm.sh` and committed. The diff is doc comments only: the
`Result<Vec<u8>, JsValue>` return does not change the emitted TypeScript
signature (`exportGlbFromMeshes` already returns `Result` and types as
`Uint8Array`) nor the `InitOutput` low-level entry, so no TypeScript
caller's types move.

Evidence: `a_degenerate_only_mesh_fails_closed_instead_of_shipping_an_invalid_dae`
(asserts the mixed state - a registered `<effect>`, no `<geometry>`, no
`<node>`, an empty `<library_geometries>` - then the refusal),
`an_empty_mesh_set_fails_closed`,
`try_collada_matches_the_infallible_path_on_real_geometry`, and
`kmz_from_an_empty_mesh_set_fails_closed_instead_of_shipping_an_empty_model`,
which first pins the tiny archive that used to ship.

`/simplify` on the branch: the infallible `export_collada_from_meshes` and
`export_kmz_collada_from_meshes` are deleted, not kept beside the `try_`
twins. This change moved their only production caller (the wasm KMZ binding)
to the fail-closed path, which left them as one-line wrappers with no caller
outside tests and an open route back to shipping a schema-invalid document as
success for anyone who reached for the shorter name. Supersede means delete
(AGENTS.md); the removal is `function_missing` against the published
ifc-lite-export 11.0.0 and rides the pending 12.0.0 major, recorded in
rust-major-offset.json. With one path left, `build_dae` IS
`try_export_collada_from_meshes`, the tuple return and the two path-equivalence
tests go, and the two KMZ packers become one `pack_kmz(model, href, opts)`.
The API-surface snapshot is regenerated (two names out, two in), and a
`@ifc-lite/wasm` patch changeset records the two behaviour changes the `.d.ts`
cannot express: `exportKmzFromMeshes` now throws where it returned an empty
archive, and `exportGlbFromMeshes` refuses one more input.
