+++ packages/wasm/pkg/ifc-lite.d.ts
@@ -399,11 +399,13 @@ export class IfcAPI {
      * per-mesh local). The caller passes exactly the meshes it wants emitted.
      *
      * Fails CLOSED: if the declared vertex/index counts run past the flattened
-     * `positions` / `indices`, there are fewer `index_counts` than meshes, or `normals`
-     * is empty or too short to cover every vertex, this throws an `Error` whose message
-     * starts with `MALFORMED_MESH_INPUT` — instead of silently emitting a GLB with those
-     * meshes dropped. (The viewer always passes fully-backed, normal-covered arrays, so
-     * this only fires on a caller bug.)
+     * `positions` / `indices`, there are fewer `index_counts` than meshes, `normals`
+     * is empty or too short to cover every vertex, or an index names a vertex its own
+     * mesh does not have (glTF 2.0 3.7.2.1), this throws an `Error` whose message
+     * starts with `MALFORMED_MESH_INPUT` — instead of silently emitting a GLB with
+     * those meshes dropped, or one carrying the out-of-range index straight into the
+     * BIN chunk. (The viewer always passes fully-backed, normal-covered, in-range
+     * arrays, so this only fires on a caller bug.)
      */
     exportGlbFromMeshes(positions: Float32Array, normals: Float32Array, indices: Uint32Array, vertex_counts: Uint32Array, index_counts: Uint32Array, colors: Float32Array, origins: Float64Array, express_ids: Uint32Array, include_metadata: boolean, lit?: boolean | null, emissive?: boolean | null): Uint8Array;
     /**
@@ -482,6 +484,14 @@ export class IfcAPI {
      * `altitude_mode` (`"clampToGround"` default ⇒ rest on terrain, ignoring
      * `altitude`; `"absolute"` ⇒ place at `altitude` metres MSL) selects the
      * KML vertical placement (#1427).
+     *
+     * Fails CLOSED: when no triangle survives (an empty visible set, or meshes
+     * whose only triangles are degenerate and collapse in the vertex dedup)
+     * this throws an `Error` whose message starts with `NO_RENDER_GEOMETRY`;
+     * a declared vertex/index count running past its buffer throws
+     * `MALFORMED_MESH_INPUT`, as `exportGlbFromMeshes` does. Both used to ship
+     * as a small "successful" archive: the first around a COLLADA document
+     * the 1.4.1 schema rejects, the second with every later mesh missing.
      */
     exportKmzFromMeshes(positions: Float32Array, normals: Float32Array, indices: Uint32Array, vertex_counts: Uint32Array, index_counts: Uint32Array, colors: Float32Array, origins: Float64Array, latitude: number, longitude: number, altitude: number, x_axis_abscissa: number | null | undefined, x_axis_ordinate: number | null | undefined, name: string, altitude_mode?: string | null): Uint8Array;
     /**
